+++ labfreed/pac_id/pac_id.py
@@ -95,12 +95,13 @@ class PAC_ID(LabFREED_BaseModel):
     
     @model_validator(mode="after")
     def _validate_issuer(self):
-        if not re.fullmatch(_domain_name_pattern, self.issuer):
+        is_pac_li = self.issuer == 'LI' # check if issuer is "PAC.LI"
+        if not (re.fullmatch(_domain_name_pattern, self.issuer) or is_pac_li ):
             self._add_validation_message(
                     source="PAC-ID",
                     level = ValidationMsgLevel.ERROR,
                     highlight_pattern=self.issuer,
-                    msg="Issuer must be a valid domain name. "
+                    msg="Issuer must be a valid domain name or 'LI'. "
                 )
          
         # recommendation that A-Z, 0-9, -, and . should be used